# models/patient.go
package models

type Patient struct {
	ID        uint   `json:"id" gorm:"primaryKey"`
	FirstName string `json:"first_name"`
	LastName  string `json:"last_name"`
	Age       int    `json:"age"`
	DoctorID  uint   `json:"doctor_id"`
}