# models/user.go
package models

type User struct {
	Id       uint   `json:"id"`
	Username string `json:"username"`
	Email    string `json:"email" gorm:"unique"`
	Password []byte `json:""`
	Role     string `json:"role"`
}